Third commit- Added Modal to display the list of ingredients added.
New Files- Modal.js component was added to display the list of ingredients,
grand total and ask whether user wants to checkout or not, Modal.module.css
for adding styling to the Modal component, OrderSummary.js component generates
the list of the ingredients dynamically.

Modifications- Added Modal component to the BurgerBuilder component for rendering.